Removed hard coded debug messages for Derby database in GenericJdbcDao.
This created an undesired hard coding to Derby (was important to have
this to print debug messages to determine the correct spring xml
configuration but this is no longer needed and this dependency is now
removed). Though, purposefully keep the commented out code in there for
future consideration. For instance, it's possible we'd want a plug-in
system for advanced per-database implementation messages as it was very
helpful to see the configuration details of the sender and receiver when
they started up in the case that there was a mis-configuration.